+++ CHANGELOG.md
@@ -9,13 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Update documentation for bisection.py
+- Add tests and function body for bisection.py
+- Add tests and function body for bisection_find_roots.py
+- update dependencies in pyproject.toml
+- add evironment.yml
+- Add tests and function body for newton1d.py
+
+## [0.0.1] - (2026-10-01)
+
 - Update code of conduct enforcement contact
 - Add documentation for bisection.py
 - Add documentation for newton1d.py
 - Add documentation for hybrid.py
 - Add documentation for plot_root.py
 - reformat .py files
-
-## [0.1.0] - (1979-01-01)
 
 - First release